requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==4.2.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
